feat(phai): replace default RunViewer layout with RunSurface compound
Rename the RunViewer compound to RunSurface and remove the baked-in DefaultRunViewerLayout, so consumers compose their own layout from RunSurface.Root + slots instead of the library dictating one. The prompt and composer slots fold into a single input-region slot (RunSurface.Composer) that owns prompt-vs-composer precedence and the null-bootstrap gate; consumers pass only the composer UI as children.

Ships a prepackaged, lazy ReadonlyRunSurface (api/readableRun) for the common read-only inbox embed, and exposes the eager RunSurface compound (api/runSurface) for custom layouts. The tasks runner scene composes its own live-composer layout; the three inbox embeds drop in ReadonlyRunSurface.
