+++ frontend/src/scenes/inbox/components/detail/AgentRunDetail.tsx
@@ -21,7 +21,7 @@ import { SANDBOX_BIND_TASK_PARAM } from 'scenes/max/maxLogic'
 import { urls } from 'scenes/urls'
 
 import { isTerminalRunStatus } from 'products/posthog_ai/frontend/api/logics'
-import { RunViewer } from 'products/posthog_ai/frontend/api/run'
+import { ReadonlyRunSurface } from 'products/posthog_ai/frontend/api/readableRun'
 import { Task, TaskRunStatus } from 'products/posthog_ai/frontend/types/taskTypes'
 
 import { inboxReportDetailLogic } from '../../logics/inboxReportDetailLogic'
@@ -193,7 +193,7 @@ function TaskLogBody({
         return (
             <div className="h-[calc(100dvh-22rem)] min-h-[420px] w-full overflow-hidden rounded border border-primary bg-surface-primary">
                 {/* In-progress runs stream live; terminal runs show the static replay. */}
-                <RunViewer
+                <ReadonlyRunSurface
                     taskId={task.id}
                     runId={runId}
                     interaction={replayOnly ? 'read-only' : 'live'}
@@ -250,7 +250,7 @@ export function OpenTaskButton({ taskId, runStatus }: { taskId: string; runStatu
 
 /**
  * Inline "Task log": the selected linked task's agent transcript, rendered with the shared
- * `RunViewer` — live for an in-progress run, static replay once terminal. A `LemonSelect`
+ * `ReadonlyRunSurface` — live for an in-progress run, static replay once terminal. A `LemonSelect`
  * switches between linked tasks (research / implementation) when there's more than one; "Open task"
  * continues the task in a new PostHog AI chat. Mirrors desktop `AgentRunDetail`'s Task-log section.
  */
@@ -301,7 +301,7 @@ function TaskLogSection({ report }: { report: SignalReport }): JSX.Element {
 
 /**
  * Agent run detail body. Shows the run state strip + output state, the linked run's agent transcript
- * inline (`TaskLogSection`, via the shared `RunViewer`), and contributing evidence.
+ * inline (`TaskLogSection`, via the shared `ReadonlyRunSurface`), and contributing evidence.
  * Mirrors desktop `AgentRunDetail`.
  */
 export function AgentRunDetail({ report }: { report: SignalReport }): JSX.Element {
