+++ products/posthog_ai/frontend/api/primitives.ts
@@ -1,8 +1,9 @@
 // Tier 2 — compound primitives & presenters: the broad presentational bucket for custom layouts and
 // bespoke threads. The Composer compound, the Thread compound + atoms, the prepackaged `ThreadView`,
 // message presenters, the shared `RunLogSkeleton` loader, activity primitives, and the
-// permission/question/resource surfaces. Use when the prepackaged `RunViewer` (Tier 1, ./run) doesn't fit
-// and you need to compose the pieces yourself — e.g. a compact inline thread via `Thread.*` atoms.
+// permission/question/resource surfaces. Use when the prepackaged `ReadonlyRunSurface` (Tier 1,
+// ./readableRun) doesn't fit and you need to compose the pieces yourself — e.g. a compact inline thread via
+// `Thread.*` atoms.
 //
 // (This is intentionally the wide presentational module; it can be subdivided later — e.g. thread.ts /
 // composer.ts — if a consumer needs finer code-splitting.)
@@ -10,7 +11,6 @@
 // Part of the `products/posthog_ai/frontend/api/<module>` public surface — import from here, not from
 // deep `../components/*` paths. See ../README.md for the tier model and ../AGENTS.md for the coupling rule.
 
-// --- Composer (Radix-style compound) ---
 export { Composer } from '../components/composer/Composer'
 export type {
     ComposerRootProps,
@@ -19,7 +19,6 @@ export type {
     ComposerSubmitProps,
 } from '../components/composer/Composer'
 
-// --- Thread + message presenters ---
 // `Thread` is the Radix-style compound (Root + Message/Markdown/Reasoning/Failure/Activity/ToolCall
 // atoms); `ThreadView` is the prepackaged virtualized presenter (also `Thread.Root`).
 export { Thread } from '../components/Thread'
@@ -30,8 +29,9 @@ export { ReasoningAnswer } from '../messages/ReasoningAnswer'
 export type { ReasoningAnswerProps } from '../messages/ReasoningAnswer'
 export { AssistantFailureMessage } from '../messages/AssistantFailureMessage'
 export { RunLogSkeleton } from '../components/RunLogSkeleton'
+export { QueuedMessageList } from '../components/QueuedMessageList'
+export type { QueuedMessageListProps } from '../components/QueuedMessageList'
 
-// --- Activity primitives ---
 export {
     Activity,
     ActivityDetails,
@@ -44,7 +44,6 @@ export {
 export type { ActivityStatus } from '../components/ActivityPrimitives'
 export { RunActivity } from '../components/RunActivity'
 
-// --- Permission / question / resource surfaces ---
 export { PermissionInput } from '../components/PermissionInput'
 export { QuestionInput } from '../components/QuestionInput'
 export { ResourcesBar } from '../components/ResourcesBar'
